pyproject.toml
[project]
name = "av-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "groq>=1.0.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
